installation requires distribute

+++ trueskilltests.py
@@ -43,7 +43,7 @@ def compare_ratings():
 @suite.test
 def rating_to_number():
     assert int(Rating(1, 2)) == 1
-    assert long(Rating(1, 2)) == 1L
+    assert long(Rating(1, 2)) == long(1)
     assert float(Rating(1.1, 2)) == 1.1
     assert complex(Rating(1.2, 2)) == 1.2+0j
